Implement deployment metrics

+++ pkg/util/util.go
@@ -5,6 +5,6 @@ import (
 	"math/rand"
 )
 
-func RandBetween(max, min int) float64 {
+func RandBetween(max, min float64) float64 {
 	return math.Floor(float64(min) + rand.Float64()*float64(max-min))
 }
